Refactor code and add implementation for ShopService and TagService
